requirements_stt.txt
av
numpy
pydub
streamlit
twilio
openai
gtts
speechrecognition
pyaudio